bug修复

+++ manager/src/main/java/com/platon/browser/dto/ticket/VoteTicket.java
@@ -2,11 +2,13 @@ package com.platon.browser.dto.ticket;
 
 import lombok.Data;
 
+import java.math.BigInteger;
+
 @Data
 public class VoteTicket {
     private Long BlockNumber;
     private String CandidateId;
-    private Long Deposit;
+    private BigInteger Deposit;
     private String Owner;
     private Long RBlockNumber;
     private Long State;
